Simplify options of the Node.js HTTP client
This simplifies the options of the client a bit, which makes the code in transports much more readable. It also exports the client, in preparation for improved test coverage.
